split operations into separate file, refactored operations/parser code
