+++ lib/parser.py
@@ -1,194 +1,10 @@
-# future must be first
-from __future__ import division
-import copy
-import numpy as np
-import operator
-
 from pyparsing import alphanums, nums, oneOf, opAssoc, operatorPrecedence,\
          CaselessLiteral, Combine, Forward, Keyword, Literal, MatchFirst,\
          OneOrMore, Optional, ParseException, Regex, Word, ZeroOrMore
 
 from lib.exceptions import ParseError
-
-
-def operator_operands(tokenlist):
-    "generator to extract operators and operands in pairs"
-    it = iter(tokenlist)
-    while 1:
-        try:
-            yield (it.next(), it.next())
-        except StopIteration:
-            break
-
-
-class EvalConstant(object):
-    """
-    Class to evaluate a parsed constant or variable
-    """
-    def __init__(self, tokens):
-        self.value = tokens[0]
-
-    def _eval(self, row):
-        try:
-            np.float64(self.value)
-            return self.value
-        except ValueError:
-            return row[self.value]
-
-
-class EvalSignOp(object):
-    """
-    Class to evaluate expressions with a leading + or - sign
-    """
-
-    def __init__(self, tokens):
-        self.sign, self.value = tokens[0]
-
-    def _eval(self, row):
-        mult = {'+':1, '-':-1}[self.sign]
-        return mult * self.value._eval(row)
-
-
-class EvalExpOp(object):
-    """
-    Class to evaluate exponentiation expressions
-    """
-
-    def __init__(self, tokens):
-        self.value = tokens[0]
-
-    def _eval(self, row):
-        _exp = self.value[0]._eval(row)
-        for op, val in operator_operands(self.value[1:]):
-            try:
-                #print 'performing: %s %s ...' % (_sum, op)
-                _exp **= np.float64(val._eval(row))
-                #print "returning prod: %s" % _sum
-                if np.isinf(_exp):
-                    return np.nan
-            except (ValueError, ZeroDivisionError):
-                #print "returning error: %s" % np.nan
-                return np.nan
-        #print "returning prod: %s" % np.nan
-        return _exp
-
-
-class EvalMultOp(object):
-    """
-    Class to evaluate addition and subtraction expressions
-    """
-
-    def __init__(self, tokens):
-        self.value = tokens[0]
-
-    def _eval(self, row):
-        try:
-            _mult = np.float64(self.value[0]._eval(row))
-            for op, val in operator_operands(self.value[1:]):
-                    if op == '*':
-                        _mult *= np.float64(val._eval(row))
-                    if op == '/':
-                        _mult /= np.float64(val._eval(row))
-                    if np.isinf(_mult):
-                        return np.nan
-        except (ValueError, ZeroDivisionError):
-            return np.nan
-        return _mult
-
-
-class EvalPlusOp(object):
-    """
-    Class to evaluate addition and subtraction expressions
-    """
-
-    def __init__(self, tokens):
-        self.value = tokens[0]
-
-    def _eval(self, row):
-        try:
-            _sum = np.float64(self.value[0]._eval(row))
-            for op, val in operator_operands(self.value[1:]):
-                    if op == '+':
-                        _sum += np.float64(val._eval(row))
-                    if op == '-':
-                        _sum -= np.float64(val._eval(row))
-                    if np.isinf(_sum):
-                        return np.nan
-        except (ValueError, ZeroDivisionError):
-            return np.nan
-        return _sum
-
-
-class EvalComparisonOp(object):
-    """
-    Class to evaluate comparison expressions
-    """
-
-    opMap = {
-        "<" : lambda a,b : a < b,
-        "<=" : lambda a,b : a <= b,
-        ">" : lambda a,b : a > b,
-        ">=" : lambda a,b : a >= b,
-        "!=" : lambda a,b : a != b,
-        "=" : lambda a,b : a == b,
-    }
-
-    def __init__(self, tokens):
-        self.value = tokens[0]
-
-    def _eval(self, row):
-        val1 = np.float64(self.value[0]._eval(row))
-        for op, val in operator_operands(self.value[1:]):
-            fn = EvalComparisonOp.opMap[op]
-            val2 = np.float64(val._eval(row))
-            if not fn(val1, val2):
-                break
-            val1 = val2
-        else:
-            return True
-        return False
-
-
-class EvalNotOp(object):
-    """
-    Class to evaluate not expressions
-    """
-
-    def __init__(self, tokens):
-        self.value = tokens[0][1]
-
-    def _eval(self, row):
-        return not self.value._eval(row)
-
-
-class EvalAndOp(object):
-    """
-    Class to evaluate and expressions
-    """
-
-    def __init__(self, tokens):
-        self.value = tokens[0]
-
-    def _eval(self, row):
-        _conj = np.bool_(self.value[0]._eval(row))
-        for op, val in operator_operands(self.value[1:]):
-            _conj = _conj and np.bool_(val._eval(row))
-        return _conj
-
-
-class EvalOrOp(object):
-    """
-    Class to evaluate or expressions
-    """
-
-    def __init__(self, tokens):
-        self.value = tokens[0]
-
-    def _eval(self, row):
-        _disj = np.bool_(self.value[0]._eval(row))
-        for op, val in operator_operands(self.value[1:]):
-            _disj = _disj or np.bool_(val._eval(row))
-        return _disj
+from lib.operations import EvalAndOp, EvalComparisonOp, EvalConstant,\
+         EvalExpOp, EvalMultOp, EvalNotOp, EvalOrOp, EvalPlusOp, EvalSignOp
 
 
 class Parser(object):
@@ -227,7 +43,7 @@ class Parser(object):
         if self.bnf:
             return self.bnf
 
-        # define literals
+        # literals
         exp_op = Literal('^')
         sign_op = oneOf('+ -')
         mult_op = oneOf('* /')
@@ -235,7 +51,9 @@ class Parser(object):
         not_op = CaselessLiteral('not')
         and_op = CaselessLiteral('and')
         or_op = CaselessLiteral('or')
+        comparison_op = oneOf("< <= > >= != =")
 
+        # operands
         integer = Word(nums)
         real = Combine(Word(nums) + '.' + Word(nums))
         reserved_words = MatchFirst(map(Keyword, self.reserved_words))
@@ -243,7 +61,7 @@ class Parser(object):
         operand = real | integer | variable
         operand.setParseAction(EvalConstant)
 
-        # define compound expressions
+        # compound expressions
         arith_expr = operatorPrecedence(operand, [
             (exp_op, 2, opAssoc.RIGHT, EvalExpOp),
             (sign_op, 1, opAssoc.RIGHT, EvalSignOp),
@@ -251,11 +69,9 @@ class Parser(object):
             (plus_op, 2, opAssoc.LEFT, EvalPlusOp),
         ])
 
-        comparison_op = oneOf("< <= > >= != =")
         comp_expr = operatorPrecedence(arith_expr, [
             (comparison_op, 2, opAssoc.LEFT, EvalComparisonOp),
         ])
-
         prop_expr = operatorPrecedence(comp_expr, [
             (not_op, 1, opAssoc.RIGHT, EvalNotOp),
             (and_op, 2, opAssoc.LEFT, EvalAndOp),
